refactor: moved files for consistency
