+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/DefaultGitHubInstallationTokenService.java
@@ -12,7 +12,7 @@ import org.kinotic.github.api.model.ProjectGitHubRepoLink;
 import org.kinotic.github.api.services.GitHubAppInstallationService;
 import org.kinotic.github.api.services.GitHubInstallationTokenService;
 import org.kinotic.github.api.services.ProjectGitHubRepoService;
-import org.kinotic.github.internal.client.GitHubInstallationTokenCache;
+import org.kinotic.github.internal.api.services.client.GitHubInstallationTokenCache;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.CompletableFuture;
