+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/DefaultProjectGitHubRepoService.java
@@ -9,8 +9,8 @@ import org.kinotic.github.api.model.GitHubAppInstallation;
 import org.kinotic.github.api.model.ProjectGitHubRepoLink;
 import org.kinotic.github.api.services.GitHubAppInstallationService;
 import org.kinotic.github.api.services.ProjectGitHubRepoService;
-import org.kinotic.github.internal.client.GitHubApiClient;
-import org.kinotic.github.internal.client.GitHubInstallationTokenCache;
+import org.kinotic.github.internal.api.services.client.GitHubApiClient;
+import org.kinotic.github.internal.api.services.client.GitHubInstallationTokenCache;
 import org.kinotic.os.internal.api.services.AbstractCrudService;
 import org.kinotic.os.internal.api.services.CrudServiceTemplate;
 import org.springframework.stereotype.Component;
