+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/GitHubAppJwtFactory.java
@@ -1,4 +1,4 @@
-package org.kinotic.github.internal.client;
+package org.kinotic.github.internal.api.services.client;
 
 import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
